% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{agri_studies}
\alias{agri_studies}
\title{Agribusiness studies survey}
\format{
A data frame with 53 rows (participants) and 42 columns (questions):
\describe{
\item{columns 1-39}{statements about agribusiness studies}
\item{columns 31-39}{personal information}
}
}
\source{
Juliette LE COLLONNIER and Lou ROBERT, students at l'Institut Agro Rennes-Angers
}
\usage{
agri_studies
}
\description{
These data were collected after a Q-method-like survey on students' expectations of agribusiness studies.
Participants had to rank how much they agreed with 38 statements about possible benefits from agribusiness studies; then, they were asked personal questions.
}
\examples{
data(agri_studies)
agri_studies[1:5, c(1,2,40,41)]
}
\keyword{datasets}
