% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{assert_choice}
\alias{assert_choice}
\title{Assert that x is one of the allowed choices}
\usage{
assert_choice(x, choices, arg_name)
}
\arguments{
\item{x}{Object to test.}

\item{choices}{Allowed values.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Assert that x is one of the allowed choices
}
\keyword{internal}
