% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{assert_column_index}
\alias{assert_column_index}
\title{Assert that a column index is valid}
\usage{
assert_column_index(x, n, arg_name = "num.var")
}
\arguments{
\item{x}{Column index.}

\item{n}{Total number of columns.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Assert that a column index is valid
}
\keyword{internal}
