% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{assert_data_frame}
\alias{assert_data_frame}
\title{Assert that x is a data frame}
\usage{
assert_data_frame(x, arg_name = "dataset")
}
\arguments{
\item{x}{Object to test.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Assert that x is a data frame
}
\keyword{internal}
