% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{assert_index_range}
\alias{assert_index_range}
\title{Assert that a column index range is valid}
\usage{
assert_index_range(first, last, n)
}
\arguments{
\item{first}{First index.}

\item{last}{Last index.}

\item{n}{Total number of columns.}
}
\description{
Assert that a column index range is valid
}
\keyword{internal}
