% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{assert_positive_integerish}
\alias{assert_positive_integerish}
\title{Assert that x is a positive integer-like scalar}
\usage{
assert_positive_integerish(x, arg_name, allow_zero = FALSE)
}
\arguments{
\item{x}{Object to test.}

\item{arg_name}{Argument name for error messages.}

\item{allow_zero}{Whether zero is accepted.}
}
\description{
Assert that x is a positive integer-like scalar
}
\keyword{internal}
