% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{assert_proportion}
\alias{assert_proportion}
\title{Assert that x is a proportion in (0, 1]}
\usage{
assert_proportion(x, arg_name)
}
\arguments{
\item{x}{A numeric scalar.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Assert that x is a proportion in (0, 1]
}
\keyword{internal}
