% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{assert_single_logical}
\alias{assert_single_logical}
\title{Assert that x is a single non-missing logical}
\usage{
assert_single_logical(x, arg_name)
}
\arguments{
\item{x}{Object to test.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Assert that x is a single non-missing logical
}
\keyword{internal}
