% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beard.R
\docType{data}
\name{beard}
\alias{beard}
\title{Beards data}
\format{
\subsection{\code{beard}}{

A contingency table (data frame) with 8 rows and 337 columns: rows are the types of beard, columns are the words used to describe them at least once.
}
}
\source{
Applied mathematics department, Institut Agro Rennes-Angers
}
\usage{
beard
}
\description{
This data refers to 8 types of beard.
Each beard was evaluated, in a few words, by 62 panelists (except beard 8 which only had 60 evaluations) in a sensometrics experiment.
}
\examples{
data(beard)
FactoMineR::descfreq(beard)

}
\keyword{datasets}
