% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{beard_wide}
\alias{beard_wide}
\title{Beards data}
\format{
\subsection{\code{beard_wide}}{

A data frame with 8 rows and 24 columns: rows are the types of beards, columns are the panelists' opinions.
}
}
\source{
Applied mathematics department, Institut Agro Rennes-Angers
}
\usage{
beard_wide
}
\description{
This data refers to 8 types of beard. It comes from a subset of the original "beard" dataset.
Each beard was evaluated, in a few words, by 24 panelists in a sensometrics experiment.
}
\examples{

data(beard_wide)
beard_wide

}
\keyword{datasets}
