% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{boss}
\alias{boss}
\title{Ideal boss survey}
\format{
A data frame with 73 rows (participants) and 39 columns (questions):
\describe{
\item{columns 1-30}{statements about the ideal boss}
\item{columns 31-39}{personal information}
}
}
\source{
Florian LECLERE and Marianne ANDRE, students at l'Institut Agro Rennes-Angers
}
\usage{
boss
}
\description{
These data were collected after a Q-method-like survey on participants' perception of an "ideal boss".
Participants had to rank how much they agreed with 30 statements about an ideal boss; then, they were asked personal questions.
}
\examples{
data(boss)
boss[1:5, c(1,2,31,33,34)]
}
\keyword{datasets}
