% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prompt.R
\name{build_standard_prompt}
\alias{build_standard_prompt}
\title{Build a standard markdown prompt}
\usage{
build_standard_prompt(introduction, request, data, conclusion = NULL)
}
\arguments{
\item{introduction}{Introduction section.}

\item{request}{Task section.}

\item{data}{Data section.}

\item{conclusion}{Optional conclusion/output-format section.}
}
\value{
A markdown prompt.
}
\description{
Build a standard markdown prompt
}
\keyword{internal}
