% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dist_mat_llm}
\alias{dist_mat_llm}
\title{LLM distance matrix}
\usage{
dist_mat_llm(ppt, n, per_miss = 0, model = "llama3", max_attempts = 3, ...)
}
\arguments{
\item{ppt}{An LLM prompt.}

\item{n}{The number of responses to be generated.}

\item{per_miss}{The probability of leaving a pairwise distance missing.}

\item{model}{Ollama model name.}

\item{max_attempts}{Maximum number of attempts used to obtain valid similarity scores.}

\item{...}{Additional arguments passed to \code{ollamar::generate}.}
}
\value{
A list containing the generated LLM responses and a distance matrix.
}
\description{
Compute a distance matrix between randomly-generated responses to an LLM prompt.
}
\examples{
\dontrun{
dist_mat_llm("Explain iris species differences", n = 5, per_miss = 0)
}
}
