% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dist_ref_llm}
\alias{dist_ref_llm}
\title{LLM response consistency}
\usage{
dist_ref_llm(ppt, ref, n, model = "llama3", max_attempts = 3, ...)
}
\arguments{
\item{ppt}{An LLM prompt.}

\item{ref}{The reference response.}

\item{n}{The number of new responses to be generated.}

\item{model}{Ollama model name.}

\item{max_attempts}{Maximum number of attempts used to obtain valid similarity scores.}

\item{...}{Additional arguments passed to \code{ollamar::generate}.}
}
\value{
A list containing generated responses and distances to the reference.
}
\description{
Compute distances between an LLM response of interest and new responses to
the same prompt.
}
\examples{
\dontrun{
dist_ref_llm("Explain iris species differences", "Reference response", n = 5)
}
}
