% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{.call_ollama_base}
\alias{.call_ollama_base}
\title{Base wrapper around ollamar::generate}
\usage{
.call_ollama_base(model, prompt, output = "df", ollama_api_options = list())
}
\arguments{
\item{model}{Model name.}

\item{prompt}{Prompt string.}

\item{output}{Output format for ollamar::generate.}

\item{ollama_api_options}{Named list of additional supported options.}
}
\value{
Result returned by ollamar::generate.
}
\description{
Base wrapper around ollamar::generate
}
\keyword{internal}
