% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{.match_llm_provider}
\alias{.match_llm_provider}
\title{Match an LLM provider}
\usage{
.match_llm_provider(provider = c("ollama", "gemini"))
}
\arguments{
\item{provider}{Provider name. Currently supports \code{"ollama"} and \code{"gemini"}.}
}
\value{
A single provider name.
}
\description{
Match an LLM provider
}
\keyword{internal}
