% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{.unit_word_qda}
\alias{.unit_word_qda}
\title{Return the wording used for product/group units in QDA prompts}
\usage{
.unit_word_qda(
  product_knowledge = c("known", "unknown"),
  capital = FALSE,
  plural = FALSE
)
}
\arguments{
\item{product_knowledge}{Either "known" or "unknown".}

\item{capital}{Logical; capitalize first letter.}

\item{plural}{Logical; return plural form.}
}
\value{
A unit label.
}
\description{
Return the wording used for product/group units in QDA prompts
}
\keyword{internal}
