% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{filter_gemini_options}
\alias{filter_gemini_options}
\title{Filter supported Gemini options}
\usage{
filter_gemini_options(extra_args)
}
\arguments{
\item{extra_args}{A named list of extra arguments.}
}
\value{
A filtered named list containing only supported Gemini options.
}
\description{
Filter supported Gemini options
}
\keyword{internal}
