% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ollama.R
\name{filter_ollama_options}
\alias{filter_ollama_options}
\title{Filter supported ollamar::generate options}
\usage{
filter_ollama_options(extra_args)
}
\arguments{
\item{extra_args}{A named list of extra arguments.}
}
\value{
A filtered named list containing only supported Ollama options.
}
\description{
Filter supported ollamar::generate options
}
\keyword{internal}
