% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{glossophobia}
\alias{glossophobia}
\title{Glossophobia survey}
\format{
A data frame with 139 rows (participants) and 41 columns (questions):
\describe{
\item{columns 1-25}{descriptions of speaking in public}
\item{columns 26-41}{personal information}
}
}
\source{
Elina BIAU and Théo LEDAIN, students at l'Institut Agro Rennes-Angers
}
\usage{
glossophobia
}
\description{
These data were collected after a Q-method-like survey on participants' feelings about speaking in public.
Participants had to rank how much they agreed with 25 descriptions of speaking in public; then, they were asked personal questions.
}
\examples{
data(glossophobia)
glossophobia[1:5, c(1,2,26,31,37)]
}
\keyword{datasets}
