% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.nail_catdes}
\alias{print.nail_catdes}
\title{Print a structured nail_catdes result}
\usage{
\method{print}{nail_catdes}(x, ...)
}
\arguments{
\item{x}{A \code{nail_catdes} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a structured nail_catdes result
}
