% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.nail_textual}
\alias{print.nail_textual}
\title{Print a structured nail_textual result}
\usage{
\method{print}{nail_textual}(x, ...)
}
\arguments{
\item{x}{A \code{nail_textual} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a structured nail_textual result
}
