% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.nail_textual_prep}
\alias{print.nail_textual_prep}
\title{Print a nail_textual_prep object}
\usage{
\method{print}{nail_textual_prep}(x, ...)
}
\arguments{
\item{x}{A \code{nail_textual_prep} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a nail_textual_prep object
}
