% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.statistical_profiles}
\alias{print.statistical_profiles}
\title{Print mechanical statistical profiles}
\usage{
\method{print}{statistical_profiles}(x, ...)
}
\arguments{
\item{x}{A \code{statistical_profiles} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print mechanical statistical profiles
}
