% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{waste}
\alias{waste}
\title{Food waste survey}
\format{
A data frame with 180 rows (participants) and 77 columns (questions).
}
\source{
Héloïse BILLES and Amélie RATEAU, students at l'Institut Agro Rennes-Angers
}
\usage{
waste
}
\description{
These data were collected after a survey on food waste, with participants describing their habits.
}
\examples{
data(waste)
waste[1:5, 1:5]
}
\keyword{datasets}
