% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{beard_cont}
\alias{beard_cont}
\title{Beard descriptions (data)}
\format{
A contingency table (data frame) with 8 rows and 337 columns:
\itemize{
\item rows are the types of beards;
\item columns are the words used at least once to describe them.
}
}
\source{
Applied mathematics department, Institut Agro Rennes-Angers
}
\usage{
beard_cont
}
\description{
These data refer to 8 types of beards.
Each beard was evaluated by 62 assessors (except beard 8 which only had 60 evaluations).
}
\examples{
data(beard_cont)
FactoMineR::descfreq(beard_cont)
}
\keyword{datasets}
