% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{beard_wide}
\alias{beard_wide}
\title{Beard descriptions (data)}
\format{
A data frame with 8 rows and 24 columns:
\itemize{
\item rows are the types of beards;
\item columns are the assessors' opinions.
}
}
\source{
Applied mathematics department, Institut Agro Rennes-Angers
}
\usage{
beard_wide
}
\description{
These data refer to 8 types of beards. They come from a subset of the original "beard" dataset.
Each beard was evaluated by 62 assessors (except beard 8 which only had 60 evaluations).
}
\examples{

data(beard_wide)
beard_wide
}
\keyword{datasets}
