% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dist_mat_llm}
\alias{dist_mat_llm}
\title{LLM Distance Matrix}
\usage{
dist_mat_llm(ppt, n, per_miss = 0)
}
\arguments{
\item{ppt}{a LLM prompt.}

\item{n}{the number of responses to generate.}

\item{per_miss}{the percentage of missing values in the final matrix (0 by default).}
}
\value{
A list containing the LLM results for each response, and the distance matrix.
}
\description{
Compute a distance matrix between randomly-generated responses to a LLM prompt.
}
\details{
The final percentage of missing values might differ from the per_miss parameter; rather than a percentage of values being turned to NA, each value has a per_miss probability of being NA.
}
\examples{
data(iris)

res_iris <- nail_catdes(iris, num.var = 5, introduction = "A study measured various parts of iris flowers from 3 different species: setosa, versicolor and virginica. I will give you the results from this study. You will have to identify what sets these flowers apart.", request = "Please explain what makes each species distinct. Also, tell me which species has the biggest flowers, and which species has the smallest.")

dist_mat_llm(res_iris$prompt, n = 5, per_miss = 0)
}
