% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dist_ref_llm}
\alias{dist_ref_llm}
\title{LLM Response Consistency}
\usage{
dist_ref_llm(ppt, ref, n)
}
\arguments{
\item{ppt}{a LLM prompt.}

\item{ref}{the reference response.}

\item{n}{the number of new responses to generate.}
}
\value{
A list containing the newly-generated prompts and a vector of their distances to the reference reponse.
}
\description{
Compute distances between a LLM response and other responses to the same prompt.
}
\examples{
data(iris)

res_iris <- nail_catdes(iris, num.var = 5, introduction = "A study measured various parts of iris flowers from 3 different species: setosa, versicolor and virginica. I will give you the results from this study. You will have to identify what sets these flowers apart.", request = "Please explain what makes each species distinct. Also, tell me which species has the biggest flowers, and which species has the smallest.")

dist_ref_llm(res_iris$prompt, res_iris$response, n = 5)
}
