% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_catdes.R
\name{nail_catdes}
\alias{nail_catdes}
\title{Analyze a categorical latent variable}
\usage{
nail_catdes(
  dataset,
  num.var,
  introduction = "",
  request =
    "Based on the results, please describe what characterizes the individuals of each group and what sets them apart from the other groups. Then, based on these characteristics, give each group a new name.",
  model = "llama3",
  isolate.groups = F,
  proba = 0.05,
  row.w = NULL
)
}
\arguments{
\item{dataset}{a data frame made up of at least one categorical variable and a set of quantitative variables and/or categorical variables.}

\item{num.var}{the number of the variable to characterized.}

\item{introduction}{the introduction for the LLM prompt.}

\item{request}{the request made to the LLM.}

\item{model}{the model name ('llama3' by default).}

\item{isolate.groups}{a boolean that indicates whether to give the LLM a single prompt, or one prompt per category. Recommended if the categorical variable has several categories.}

\item{proba}{the significance threshold considered to characterized the category (by default 0.05).}

\item{row.w}{a vector of integers corresponding to an optional row weights (by default, a vector of 1 for uniform row weights)}
}
\value{
A data frame, or a list of data frames, containing the LLM's response.
}
\description{
Generate a LLM response to analyze a categorical latent variable.
}
\details{
This function directly sends a prompt to a LLM. Therefore, to get a consistent answer, we highly recommend to customize the parameters introduction and request and add all relevant information on your data for the LLM. We also recommend renaming the columns to clear, unshortened and unambiguous names.

Additionally, if isolate.groups = TRUE, you will need an introduction and a request that take into account the fact that only one group is analyzed at a time.
}
\examples{
data(iris)

res = nail_catdes(iris, num.var = 5, introduction = "A study measured various parts of iris flowers from 3 different species: setosa, versicolor and virginica. I will give you the results from this study. You will have to identify what sets these flowers apart.", request = "Please explain what makes each species distinct. Also, tell me which species has the biggest flowers, and which species has the smallest.")

cat(res$response)
}
