% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_condes.R
\name{nail_condes}
\alias{nail_condes}
\title{Interpret a continuous latent variable}
\usage{
nail_condes(
  dataset,
  num.var,
  introduction = "",
  request =
    "Please explain what differentiates individuals from both sides of the scale. Then give a name to the scale, and briefly explain why you chose that name.",
  model = "llama3",
  quanti.threshold = 0,
  quanti.mod = c("Significantly above average", "Significantly below average", "Average"),
  weights = NULL,
  proba = 0.05,
  generate = T
)
}
\arguments{
\item{dataset}{a data frame made up of at least one quantitative variable and a set of quantitative variables and/or categorical variables.}

\item{num.var}{the number of the variable to be characterized.}

\item{introduction}{the introduction for the LLM prompt.}

\item{request}{the request made to the LLM.}

\item{model}{the model name ('llama3' by default).}

\item{quanti.threshold}{the threshold above (resp. below) which a scaled variable is considered significantly above (resp.below) the average. Used when converting continuous variables to categorical.}

\item{quanti.mod}{a vector of the 3 possible modalities for continuous variables converted to categorical according to the threshold. Default is "above average", "below average" and "average".}

\item{weights}{weights for the individuals (see \code{\link[FactoMineR:condes]{FactoMineR::condes()}}).}

\item{proba}{the significance threshold considered to characterize the category (by default 0.05).}
}
\value{
A data frame containing the LLM's response.
}
\description{
Generate an LLM response to analyze a continuous latent variable.
}
\details{
This function directly sends a prompt to an LLM. Therefore, to get a consistent answer, we highly recommend to customize the parameters introduction and request and add all relevant information on your data for the LLM. We also recommend renaming the columns with clear, unshortened and unambiguous names.
}
\examples{
library(FactoMineR)
data(decathlon)

names(decathlon) <- c('Time taken to complete the 100m', 'Distance reached for the long jump', 'Distance reached for the shot put', 'Height reached for the high jump',  'Time taken to complete the 400m', 'Time taken to complete the 110m hurdle', 'Distance reached for the discus', 'Height reached for the pole vault', 'Distance reached for the javeline', 'Time taken to complete the 1500 m', 'Rank/Counter-performance indicator', 'Points', 'Competition')

res_pca_deca <- FactoMineR::PCA(decathlon, quanti.sup = 11:12, quali.sup = 13)
deca_work <- res_pca_deca$ind$coord |> as.data.frame()
deca_work <- deca_work[,1] |> cbind(decathlon)

res_deca <- nail_condes(deca_work, 1, quanti.threshold = 1, quanti.mod = c('High', 'Low', 'Average'), introduction = "A study was led on athletes participating to a decathlon event. Their performance was assessed on each part of the decathlon, and they were all placed on a unidimensional scale.")

cat(res_deca$response)
}
