% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_condes.R
\name{nail_condes}
\alias{nail_condes}
\title{Analyze a continuous latent variable}
\usage{
nail_condes(
  dataset,
  num.var,
  introduction = "",
  request =
    "Please explain what differentiates individuals from both sides of the scale. Then give a name to the scale, and briefly explain why you chose that name.",
  model = "llama3",
  recode = 1,
  weights = NULL,
  proba = 0.05
)
}
\arguments{
\item{dataset}{a data frame made up of at least one quantitative variable and a set of quantitative variables and/or categorical variables.}

\item{num.var}{the number of the variable to characterized.}

\item{introduction}{the introduction for the LLM prompt.}

\item{request}{the request made to the LLM.}

\item{model}{the model name ('llama3' by default).}

\item{recode}{an integer specifying how continuous variables are converted to categorical, for the analysis. recode = 1 (default) converts values to "more/less than average"; recode = 2 adds the categories "significantly more/less than average" when a value is more/less than 2 standard deviations away from the average.}

\item{weights}{weights for the individuals (see \code{\link[FactoMineR:condes]{FactoMineR::condes()}}).}

\item{proba}{the significance threshold considered to characterized the category (by default 0.05).}
}
\value{
A data frame containing the LLM's response.
}
\description{
Generate a LLM response to analyze a continuous latent variable
}
\details{
This function directly sends a prompt to a LLM. Therefore, to get a consistent answer, we highly recommend to customize the parameters introduction and request and add all relevant information on your data for the LLM. We also recommend renaming the columns to clear, unshortened and unambiguous names.
}
\examples{
data(decathlon)

names(decathlon) = c('Time taken to complete the 100m', 'Distance reached for the long jump', 'Distance reached for the shot put', 'Height reached for the high jump',  'Time taken to complete the 400m', 'Time taken to complete the 110m hurdle', 'Distance reached for the discus', 'Height reached for the pole vault', 'Distance reached for the javeline', 'Time taken to complete the 1500 m', 'Rank', 'Points', 'Competition')

res1 = nail_condes(deca_work, 1, recode = 2, introduction = "A study was led on athletes participating to a decathlon event. Their performance was assessed on each part of the decathlon, and they were all placed on a unidimensional scale. You will analyze what sets apart individuals from either side of the scale. If an event is listed as both below (or above) and significantly below (or above) average, consider it significantly so.")

cat(res1$response)
}
