% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_descfreq.R
\name{nail_descfreq}
\alias{nail_descfreq}
\title{Interpret the rows of a contingency table}
\usage{
nail_descfreq(
  dataset,
  introduction = "",
  request = "Based on the results, please describe what makes each row unique.",
  model = "llama3",
  isolate.groups = F,
  by.quali = NULL,
  proba = 0.05,
  generate = T
)
}
\arguments{
\item{dataset}{a data frame corresponding to a contingency table.}

\item{introduction}{the introduction for the LLM prompt.}

\item{request}{the request made to the LLM.}

\item{model}{the model name ('llama3' by default).}

\item{isolate.groups}{a boolean that indicates whether to give the LLM a single prompt, or one prompt per row. Recommended if the contingency table has a great number of rows.}

\item{by.quali}{a factor used to merge the data from different rows of the contingency table; by default NULL and each row is characterized.}

\item{proba}{the significance threshold considered to characterize the category (by default 0.05).}
}
\value{
A data frame, or a list of data frames (if isolate.groups = TRUE), containing the LLM's response.
}
\description{
Describes the rows of a contingency table. For each row, this description is based on the columns of the contingency table that are significantly related to it.
}
\details{
This function directly sends a prompt to an LLM. Therefore, to get a consistent answer, we highly recommend to customize the parameters introduction and request and add all relevant information on your data for the LLM.

Additionally, if isolate.groups = TRUE, you will need an introduction and a request that take into account the fact that only one group is analyzed at a time.
}
\examples{
### Example 1: beard dataset ###

data(beard_cont)

res_beard <- nail_descfreq(beard_cont, introduction = 'A survey was conducted about beards and 8 types of beards were described. I will give you the results for one type of beard.', request = 'Please give a name to this beard and summarise what makes this beard unique.', isolate.groups = T, generate = F)

res_beard$prompt[1]
res_beard$prompt[2]

res_beard <- nail_descfreq(beard_cont, introduction = 'A survey was conducted about beards and 8 types of beards were described. In the data that follows, beards are named B1 to B8.', request = 'Please give a name to each beard and summarise what makes this beard unique.')
cat(res_beard$response)

\dontrun{
text <- res_beard$response
titles <- stringr::str_extract_all(text, "\\\\*\\\\*B[0-9]+: [^\\\\*\\\\*]+\\\\*\\\\*")[[1]]

titles

# for the following code to work, the response must have the beards' new names with this format: **B1: The Nice beard**, etc.

titles <- stringr::str_replace_all(titles, "\\\\*\\\\*", "")  # remove asterisks
names <- stringr::str_extract(titles, ": .+")
names <- stringr::str_replace_all(names, ": ", "")  # remove the colon and space

rownames(beard_cont) <- names

library(FactoMineR)

res_ca_beard <- CA(beard_cont, graph = F)
plot.CA(res_ca_beard, invisible = "col")
}


### Example 2: children dataset ###

data(children)

children <- children[1:14, 1:5] |> t() |> as.data.frame()
rownames(children) <- c('No education', 'Elementary school', 'Middle school', 'High school', 'University')

res_children <- nail_descfreq(children, introduction = 'The data used here is a contingency table that summarizes the answers given by different categories of people to the following question: "according to you, what are the reasons that can make hesitate a woman or a couple to have children?". Each row corresponds to a level of education, and columns are reasons.', request = "Please explain the main differences between more educated and less educated couples, when it comes to hesitating to have children.")

cat(res_children$response)
}
