% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_descfreq.R
\name{nail_descfreq}
\alias{nail_descfreq}
\title{Analyze a latent variable in a contingency table}
\usage{
nail_descfreq(
  dataset,
  introduction = "",
  request =
    "Based on the results, please describe what characterizes the individuals of each group. Then, based on these characteristics, give each group a new name..",
  model = "llama3",
  isolate.groups = F,
  by.quali = NULL,
  proba = 0.05
)
}
\arguments{
\item{dataset}{a data frame corresponding to a contingency table (quantitative data).}

\item{introduction}{the introduction for the LLM prompt.}

\item{request}{the request made to the LLM.}

\item{model}{the model name ('llama3' by default).}

\item{isolate.groups}{a boolean that indicates whether to give the LLM a single prompt, or one prompt per row. Recommended if the contingency table has several rows.}

\item{by.quali}{a factor used to merge the data from different rows of the contingency table; by default NULL and each row is characterized}

\item{proba}{the significance threshold considered to characterized the category (by default 0.05).}
}
\value{
A data frame, or a list of data frames, containing the LLM's response.
}
\description{
Generate a LLM response to analyze a latent variable in a contingency table.
}
