% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_descfreq.R
\name{nail_descfreq}
\alias{nail_descfreq}
\title{Analyze a latent variable in a contingency table}
\usage{
nail_descfreq(
  dataset,
  introduction = "",
  request =
    "Based on the results, please describe what characterizes the individuals of each group. Then, based on these characteristics, give each group a new name..",
  model = "llama3",
  isolate.groups = F,
  by.quali = NULL,
  proba = 0.05
)
}
\arguments{
\item{dataset}{a data frame corresponding to a contingency table (quantitative data).}

\item{introduction}{the introduction for the LLM prompt.}

\item{request}{the request made to the LLM.}

\item{model}{the model name ('llama3' by default).}

\item{isolate.groups}{a boolean that indicates whether to give the LLM a single prompt, or one prompt per row. Recommended if the contingency table has several rows.}

\item{by.quali}{a factor used to merge the data from different rows of the contingency table; by default NULL and each row is characterized}

\item{proba}{the significance threshold considered to characterized the category (by default 0.05).}
}
\value{
A data frame, or a list of data frames, containing the LLM's response.
}
\description{
Generate a LLM response to analyze a latent variable in a contingency table.
}
\details{
This function directly sends a prompt to a LLM. Therefore, to get a consistent answer, we highly recommend to customize the parameters introduction and request and add all relevant information on your data for the LLM.

Additionally, if isolate.groups = TRUE, you will need an introduction and a request that take into account the fact that only one group is analyzed at a time.
}
\examples{
data(beard)
nail_descfreq(beard, introduction = 'A survey about beards was led, and 62 participants had to describe 8 types of beards. I will give you the results for one type of beard and you will have to summarise what makes this beard unique.', request = 'Please summarise what makes this beard unique, in no more than 3 sentences', isolate.groups = T)
}
