% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nail_textual.R
\name{nail_textual}
\alias{nail_textual}
\title{Interpret a group based on answers to open-ended questions}
\usage{
nail_textual(
  dataset,
  num.var,
  num.text,
  introduction = "",
  request = NULL,
  model = "llama3",
  isolate.groups = TRUE,
  generate = FALSE
)
}
\arguments{
\item{dataset}{a data frame made up of at least one categorical variable and a textual variable.}

\item{num.var}{the index of the categorical variable to be characterized.}

\item{num.text}{the index of the textual variable that characterizes the categorical variable of interest.}

\item{introduction}{the introduction for the LLM prompt.}

\item{request}{the request made to the LLM.}

\item{model}{the model name ('llama3' by default).}

\item{isolate.groups}{a boolean that indicates whether to give the LLM a single prompt, or one prompt per category. Recommended with long catdes results.}

\item{generate}{a boolean that indicates whether to generate the LLM response. If FALSE, the function only returns the prompt.}
}
\value{
A data frame, or a list of data frames, containing the LLM's prompt and response (if generate = TRUE).
}
\description{
Generate an LLM response to analyze a categorical latent variable, based on answers to open-ended questions.
}
\details{
This function directly sends a prompt to an LLM. Therefore, to get a consistent answer, we highly recommend to customize the parameters introduction and request and add all relevant information on your data for the LLM. We also recommend renaming the columns with clear, unshortened and unambiguous names.

Additionally, if isolate.groups = TRUE, you will need an introduction and a request that take into account the fact that only one group is analyzed at a time.
}
\examples{
\dontrun{
# Processing time is often longer than ten seconds
# because the function uses a large language model.

### Example 1: Car alone survey ###
library(NaileR)
library(dplyr)
data(car_alone)

sampled_car_alone <- car_alone \%>\%
group_by(car_alone_capable_restrictive) \%>\%
dplyr::sample_frac(0.5)
sampled_car_alone <- as.data.frame(sampled_car_alone)

intro_car <- "Knowing the impact on the climate,
I have made these choices based on
the following benefits and constraints..."
intro_car <- gsub('\n', ' ', intro_car) |>
stringr::str_squish()

res_nail_textual <- nail_textual(sampled_car_alone, num.var = 1,
                                 num.text = 2,
                                 introduction = intro_car,
                                 request = NULL,
                                 model = 'llama3', isolate.groups = TRUE,
                                 generate = TRUE)
res_nail_textual[[1]]$response |> cat()
res_nail_textual[[3]]$response |> cat()
res_nail_textual[[2]]$response |> cat()
res_nail_textual[[4]]$response |> cat()

### Example 2: Atomic habits survey ###
library(NaileR)
library(dplyr)
data(atomic_habit_clust)

intro_atomic <- "These data were collected
after a survey on atomic habits: we asked
what people were prepared to change about their daily habits
to make the world a better place,
what habits they felt able to adopt,
what habits were restrictive."
intro_atomic <- gsub('\n', ' ', intro_atomic) |>
stringr::str_squish()

dta_plane <- atomic_habit_clust[,c(32,51)] \%>\%
            filter(never_plane_text != 'THAT')

sampled_dta_plane <- dta_plane \%>\%
                    group_by(clust) \%>\%
                    dplyr::sample_frac(0.75)

sampled_dta_plane <- as.data.frame(sampled_dta_plane)
summary(sampled_dta_plane)

res_nail_textual_plane <- nail_textual(sampled_dta_plane, num.var = 2,
                                      num.text = 1,
                                      introduction = intro_atomic,
                                      request = NULL,
                                      model = 'llama3',
                                      isolate.groups = TRUE,
                                      generate = TRUE)
cat(res_nail_textual_plane[[1]]$prompt)
cat(res_nail_textual_plane[[1]]$response)

cat(res_nail_textual_plane[[2]]$prompt)
cat(res_nail_textual_plane[[2]]$response)

cat(res_nail_textual_plane[[3]]$prompt)
cat(res_nail_textual_plane[[3]]$response)

res_nail_textual_plane <- nail_textual(sampled_dta_plane, num.var = 2,
                                      num.text = 1,
                                      introduction = intro_atomic,
                                      request = NULL,
                                      model = 'llama3',
                                      isolate.groups = FALSE,
                                      generate = TRUE)
cat(res_nail_textual_plane$prompt)
cat(res_nail_textual_plane$response)
}
}
