% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nutriscore}
\alias{nutriscore}
\title{Nutri-score survey}
\format{
A data frame with 112 rows (participants) and 36 columns (questions).
}
\source{
Anaëlle YANNIC and Jessie PICOT, students at l'Institut Agro Rennes-Angers
}
\usage{
nutriscore
}
\description{
These data were collected after a survey on the nutri-score. Participants were asked various questions about their views on the nutri-score, and about their eating habits.
}
\examples{
data(nutriscore)
nutriscore[1:5, c(1,8,21,30)]
}
\keyword{datasets}
