% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{sim_llm}
\alias{sim_llm}
\title{LLM Text Similarity}
\usage{
sim_llm(textA, textB)
}
\arguments{
\item{textA, textB}{two character strings}
}
\value{
An integer between 0 and 100.
}
\description{
Compute a similarity score, on a scale ranging from 0 to 100, between two character strings.
}
\details{
The similarity score is generated by a LLM. Therefore, the result might vary if the function is run several times.
}
\examples{
textA <- "Participant A was described as a nice, outgoing man, with a friendly attitude."
textB <- "Participant A was an extroverted and caring individual."

sim_llm(textA, textB)
}
